dashboard/backend/tests/unit/test_auth_service.py
"""
Unit Tests for Auth Service

Tests authentication, JWT generation, and user management.
"""
import pytest
from unittest.mock import Mock, AsyncMock, patch
from datetime import datetime, timedelta
from jose import jwt

from services.auth_service import (
    hash_password,
    verify_password,
    create_access_token,
    authenticate_user,
    get_current_user
)
from models import User
from schemas import UserCreate, LoginRequest


@pytest.mark.asyncio
class TestAuthService:
    """Test suite for auth service"""

    def test_hash_password(self):
        """Test password hashing"""
        password = "test_pwd_123"  # Shorter password (< 72 bytes)
        hashed = hash_password(password)
        
        # Verify it's hashed
        assert hashed != password
        assert hashed.startswith("$2b$")
        assert len(hashed) > 50

    def test_verify_password_correct(self):
        """Test password verification with correct password"""
        password = "test_pwd_123"  # Shorter password (< 72 bytes)
        hashed = hash_password(password)
        
        # Verify correct password
        assert verify_password(password, hashed) is True

    def test_verify_password_incorrect(self):
        """Test password verification with incorrect password"""
        password = "test_pwd_123"  # Shorter password (< 72 bytes)
        hashed = hash_password(password)
        
        # Verify incorrect password
        assert verify_password("wrong_pwd", hashed) is False

    def test_create_access_token(self):
        """Test JWT token creation"""
        data = {"sub": "testuser"}
        token = create_access_token(data)
        
        # Verify token is created
        assert token is not None
        assert isinstance(token, str)
        
        # Decode and verify
        payload = jwt.decode(token, "your-secret-key-change-in-production", algorithms=["HS256"])
        assert payload["sub"] == "testuser"
        assert "exp" in payload

    def test_create_access_token_with_expiry(self):
        """Test JWT token creation with custom expiry"""
        data = {"sub": "testuser"}
        expires = timedelta(minutes=15)
        token = create_access_token(data, expires)
        
        # Decode and verify expiry
        payload = jwt.decode(token, "your-secret-key-change-in-production", algorithms=["HS256"])
        exp_time = datetime.fromtimestamp(payload["exp"])
        
        # Should be approximately 15 minutes from now
        now = datetime.utcnow()
        delta = exp_time - now
        
        assert 14 * 60 < delta.total_seconds() < 16 * 60

    async def test_authenticate_user_success(self):
        """Test successful user authentication"""
        # Mock database session
        mock_db = AsyncMock()
        
        # Mock user
        user = User(
            id=1,
            username="testuser",
            email="test@example.com",
            hashed_password=hash_password("test_pwd_123"),  # Shorter password
            is_active=True
        )
        
        # Mock database query
        mock_result = AsyncMock()
        mock_result.scalar_one_or_none = Mock(return_value=user)  # Sync mock
        mock_db.execute = AsyncMock(return_value=mock_result)
        
        # Test authentication
        authenticated_user = await authenticate_user(mock_db, "testuser", "test_pwd_123")
        
        assert authenticated_user is not None
        assert authenticated_user.username == "testuser"

    async def test_authenticate_user_wrong_password(self):
        """Test authentication with wrong password"""
        # Mock database session
        mock_db = AsyncMock()
        
        # Mock user
        user = User(
            id=1,
            username="testuser",
            email="test@example.com",
            hashed_password=hash_password("test_pwd_123"),  # Shorter password
            is_active=True
        )
        
        # Mock database query
        mock_result = AsyncMock()
        mock_result.scalar_one_or_none = Mock(return_value=user)  # Sync mock
        mock_db.execute = AsyncMock(return_value=mock_result)
        
        # Test authentication with wrong password
        authenticated_user = await authenticate_user(mock_db, "testuser", "wrong_pwd")
        
        assert authenticated_user is None

    async def test_authenticate_user_not_found(self):
        """Test authentication with non-existent user"""
        # Mock database session
        mock_db = AsyncMock()
        
        # Mock database query returning None
        mock_result = AsyncMock()
        mock_result.scalar_one_or_none = Mock(return_value=None)  # Sync mock
        mock_db.execute = AsyncMock(return_value=mock_result)
        
        # Test authentication
        authenticated_user = await authenticate_user(mock_db, "nonexistent", "pwd123")
        
        assert authenticated_user is None

    async def test_get_current_user_success(self):
        """Test getting current user from valid token"""
        # Mock database session
        mock_db = AsyncMock()
        
        # Mock user
        user = User(
            id=1,
            username="testuser",
            email="test@example.com",
            hashed_password="hashed",
            is_active=True
        )
        
        # Mock database query
        mock_result = AsyncMock()
        mock_result.scalar_one_or_none = Mock(return_value=user)  # Sync mock
        mock_db.execute = AsyncMock(return_value=mock_result)
        
        # Create token
        token = create_access_token({"sub": "testuser"})
        
        # Mock credentials
        mock_credentials = Mock()
        mock_credentials.credentials = token
        
        # Test get current user
        current_user = await get_current_user(mock_credentials, mock_db)
        
        assert current_user is not None
        assert current_user.username == "testuser"


@pytest.mark.asyncio
class TestUserCreation:
    """Test user creation and validation"""

    async def test_create_user_password_hashing(self):
        """Test that passwords are properly hashed when creating users"""
        from services.user_service import create_user_service
        
        # Mock database
        mock_db = AsyncMock()
        
        # Mock no existing users
        mock_result = AsyncMock()
        mock_result.scalar_one_or_none.return_value = None
        mock_db.execute.return_value = mock_result
        
        # Create user
        user_data = UserCreate(
            username="newuser",
            email="new@example.com",
            password="plain_pwd",  # Shorter password
            is_active=True
        )
        
        # Mock add and commit
        mock_db.add = Mock()
        mock_db.commit = AsyncMock()
        mock_db.refresh = AsyncMock()
        
        # This would normally create the user
        # We're testing that the password gets hashed
        hashed = hash_password("plain_pwd")
        
        # Verify it's properly hashed
        assert hashed != "plain_pwd"
        assert verify_password("plain_pwd", hashed) is True


# Run tests
if __name__ == "__main__":
    pytest.main([__file__, "-v"])