dashboard/backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
asyncpg==0.30.0
pydantic==2.9.2
pydantic-settings==2.6.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
aiofiles==24.1.0
websockets==14.0
redis==5.2.0
rq==1.16.2
celery==5.4.0
flower==2.0.1
pytest==8.3.3
pytest-asyncio==0.25.0
pytest-cov==5.0.0
pytest-playwright==0.7.2
playwright==1.58.0
httpx==0.27.2
allure-pytest==2.13.5
psycopg2-binary==2.9.10
python-dotenv==1.0.1
structlog==24.4.0
locust==2.20.0